♻️ Refactor RecentsSection to server component
Split RecentsSection into server and client components to maintain
server-side data fetching while preserving client-side active state
indicator functionality.

- Create RecentsSectionClient for client-side logic
- Move data fetching to server component
- Add fetchRecentSessions for server-side data fetching
- Extract RecentSession type to separate file
- Remove unused client-side fetch function
