+++ frontend/apps/app/components/CommonLayout/GlobalNav/RecentsSection/types.ts
@@ -1,0 +1,6 @@
+export type RecentSession = {
+  id: string
+  name: string
+  created_at: string
+  project_id: string | null
+}
